+++ src/js/module/require-config.js
@@ -3,13 +3,13 @@
 require = { // eslint-disable-line no-global-assign
     'baseUrl': '/',
     'paths': {
-        'vue': 'js/require/vue.js?v=2.6.11',
-        'vue-router': 'js/require/vue-router.js?v=3.3.2',
-        'vue-i18n': 'js/require/vue-i18n.js?v=8.18.1',
-        '@fortawesome/fontawesome-svg-core': 'js/require/fontawesome-svg-core.js?v=1.2.28',
-        '@fortawesome/free-brands-svg-icons': 'js/require/free-brands-svg-icons.js?v=5.13.0',
-        '@fortawesome/free-regular-svg-icons': 'js/require/free-regular-svg-icons.js?v=5.13.0',
-        '@fortawesome/free-solid-svg-icons': 'js/require/free-solid-svg-icons.js?v=5.13.0',
-        '@fortawesome/vue-fontawesome': 'js/require/vue-fontawesome.js?v=0.1.9'
+        'vue': 'js/require/vue.js?v=2.6.12',
+        'vue-router': 'js/require/vue-router.js?v=3.4.3',
+        'vue-i18n': 'js/require/vue-i18n.js?v=8.21.1',
+        '@fortawesome/fontawesome-svg-core': 'js/require/fontawesome-svg-core.js?v=1.2.30',
+        '@fortawesome/free-brands-svg-icons': 'js/require/free-brands-svg-icons.js?v=5.14.0',
+        '@fortawesome/free-regular-svg-icons': 'js/require/free-regular-svg-icons.js?v=5.14.0',
+        '@fortawesome/free-solid-svg-icons': 'js/require/free-solid-svg-icons.js?v=5.14.0',
+        '@fortawesome/vue-fontawesome': 'js/require/vue-fontawesome.js?v=2.0.0'
     }
 };